@@ -1,12 +1,11 @@
 name: easy_anim
 description: Declarative description uses Flutter animation. Develop complex animations in a simpler way
-version: 2.1.1
-author: fenger
+version: 3.1.1
 homepage: https://github.com/fengerwoo
 repository: https://github.com/fengerwoo/easy_anim
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.1 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
